config: add `@` alias to reference `./src`
This will simplify imports and make them relative to the `./src` folder, not relative to the current path.

For example, when importing from Column the `./src/logic/board` instead of using `../logic/board`, we can now use
`@/logic/board`.

The benefit might not seem huge at this level, but
when the app grows, this makes importing
much easier, as it also allows to migrate logic
to other files without worrying about the nesting
level of the new file because having the imports
already relative to the `./src` folder.